# Use `rb_sym2str()` to retrieve String object