@Watson1978 2021-08-09
> BTW, posted an article on Hacker News about the new Oj parser.

Nice. I found the article :)
https://news.ycombinator.com/item?id=28115675